# server/ebooking-api/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "API",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Administrator/GetPartners": {
      "get": {
        "tags": [
          "Administrator"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Administrator/Add": {
      "post": {
        "tags": [
          "Administrator"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdministratorPOST"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AdministratorPOST"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AdministratorPOST"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Administrator/Delete/{id}": {
      "delete": {
        "tags": [
          "Administrator"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Administrator/Update": {
      "patch": {
        "tags": [
          "Administrator"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdministratorPATCH"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AdministratorPATCH"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AdministratorPATCH"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Administrator/Details": {
      "get": {
        "tags": [
          "Administrator"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Administrator/Accommodations": {
      "get": {
        "tags": [
          "Administrator"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Auth/refresh-token": {
      "get": {
        "tags": [
          "Auth"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Auth/facebook-login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FacebookSignInDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/FacebookSignInDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/FacebookSignInDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Auth/google-login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoogleSignInVM"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/GoogleSignInVM"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/GoogleSignInVM"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Auth/status": {
      "get": {
        "tags": [
          "Auth"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Country/Add": {
      "post": {
        "tags": [
          "Country"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CountryPOST"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CountryPOST"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CountryPOST"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Country/Get/{id}": {
      "get": {
        "tags": [
          "Country"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Country/Delete/{id}": {
      "delete": {
        "tags": [
          "Country"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Country/Update": {
      "patch": {
        "tags": [
          "Country"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CountryPATCH"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CountryPATCH"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CountryPATCH"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdministratorPATCH": {
        "type": "object",
        "additionalProperties": false
      },
      "AdministratorPOST": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string",
            "nullable": true
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "birthDate": {
            "type": "string",
            "format": "date-time"
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          },
          "image": {
            "type": "string",
            "format": "byte",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "CountryPATCH": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "maxLength": 50,
            "minLength": 5,
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CountryPOST": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "maxLength": 50,
            "minLength": 5,
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "FacebookSignInDTO": {
        "required": [
          "accessToken"
        ],
        "type": "object",
        "properties": {
          "accessToken": {
            "minLength": 1,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "GoogleSignInVM": {
        "required": [
          "accessToken",
          "idToken"
        ],
        "type": "object",
        "properties": {
          "idToken": {
            "minLength": 1,
            "type": "string"
          },
          "accessToken": {
            "minLength": 1,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "LoginDTO": {
        "required": [
          "email"
        ],
        "type": "object",
        "properties": {
          "email": {
            "minLength": 1,
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "format": "password",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}